+++ TKGame/Content/Content.mgcb
@@ -37,3 +37,15 @@
 /processorParam:TextureFormat=Color
 /build:Art/Player.png
 
+#begin Art/Weapons/IronSword.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Art/Weapons/IronSword.png
+
