+++ TKGame/Content/Weapons/Sword.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TKGame.Content.Weapons
+{
+    class Sword : Weapon
+    {
+        public override int DamageEnemy(Weapon weapon)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
